axios req for images and files
