+++ package.json
@@ -8,7 +8,9 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "axios": "^0.25.0",
     "bootstrap": "^5.1.3",
+    "file-exists": "^5.0.1",
     "file-saver": "^2.0.5",
     "gh-pages": "^3.2.3",
     "jszip": "^3.7.1",
