+++ src/components/DownloadButton.js
@@ -14,45 +14,9 @@ function DownloadPlotsButton({
 }) {
   // Handles what happens when save plot button is clicked
   const handleClick = () => {
-    FileCheck();
+    const imglist = FileCheck({ plotvars, seasons, periods, plottypes }, "png");
+    let downloadlist = imglist[0];
 
-    let imglist = [];
-    let warnings = [];
-    plotvars.map((pvar) =>
-      seasons.map((season) =>
-        periods.map((period) =>
-          plottypes.map((ptype) => {
-            // TODO: add region
-            // regions.map((region) => {
-            // TODO: when adding data download functionality, this is the section to
-            // change the beginning of the path
-            let path =
-              `${process.env.PUBLIC_URL}` +
-              "/images/" +
-              pvar.value +
-              "/" +
-              ptype.value +
-              "_" +
-              pvar.value +
-              "_" +
-              season.value +
-              "_" +
-              period.value +
-              ".png";
-
-            return [
-              imglist.push(path),
-              warnings.push([
-                pvar.label,
-                ptype.label,
-                season.label,
-                period.label,
-              ]),
-            ];
-          })
-        )
-      )
-    );
     // call zip function here
     return downloadzip(imglist, "plots");
   };
